requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyJWT
psycopg2-binary
opencv-python
Pillow
torch
torchvision
numpy
python-dotenv
sqlalchemy
passlib[bcrypt] 